requirements.txt
flask
flask-sqlalchemy
flask-security
flask-wtf
flask-nav
flask-bootstrap
marshmallow-sqlalchemy
flask-restx
flask-marshmallow==0.11.0
Flask-Continuum
sqlalchemy-utils
netaddr
netmiko==2.4.2
scrapli
scrapli[full]
napalm
pandas
dynaconf
toml
dictdiffer

